Initial commit of townsquare_spy module.
Fairly basic and not entirely well-documented, but seems to work
and includes tests. Can also be run as main to provide a simple
and practical way of testing it on actual running games.

In the future, this could be linked to the bot to allow it to
start monitoring games which are linked to, and gathering information
for auditing or possibly certain statistics (but noting that the
townsquare tool doesn't provide all the info one might want, and
isn't an altogether secure system against abuse).